+++ src/main/java/com/vtence/molecule/servers/SimpleServer.java
@@ -110,9 +110,17 @@ public class SimpleServer implements Server {
             request.remotePort(httpRequest.getClientAddress().getPort());
             request.remoteHost(httpRequest.getClientAddress().getHostName());
             request.timestamp(httpRequest.getRequestTime());
+            request.scheme(schemeOf(httpRequest));
             request.protocol(String.format("HTTP/%s.%s", httpRequest.getMajor(), httpRequest.getMinor()));
             request.secure(httpRequest.isSecure());
             request.method(httpRequest.getMethod());
+        }
+
+        private String schemeOf(org.simpleframework.http.Request httpRequest) {
+            // Prefer the scheme specified in the host header if any
+            String scheme = httpRequest.getAddress().getScheme();
+            if (scheme != null) return scheme;
+            return httpRequest.isSecure() ? "https" : "http";
         }
 
         private void readHeaders(Request request, org.simpleframework.http.Request httpRequest) {